fix(refactor): removed not used filed, transfered a couple of components to library
