--- apps/Conduit-UI/src/pages/database.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { Box } from '@mui/material';
-
-const Database: React.FC = () => {
-  return (
-    <Box sx={{ height: '100vh', display: 'flex', justifyContent: 'center', alignItems: 'center' }}>
-      Coming not so soon
-    </Box>
-  );
-};
-
-export default Database;
