+++ libraries/ui-components/package.json
@@ -43,6 +43,7 @@
     "notistack": "^2.0.3",
     "next": "^12.1.0",
     "axios": "^0.26.0",
-    "lodash": "^4.17.15"
+    "lodash": "^4.17.15",
+    "react-spring": "^9.4.3"
   }
 }
